+++ plugins/elektro_arktika_support/elektro_arktika/lrit/module_elektro_lrit_data_decoder.h
@@ -32,7 +32,7 @@ namespace elektro
 
                 // UI Stuff
                 bool hasToUpdate = false;
-                unsigned int textureID = 0;
+                intptr_t textureID = 0;
                 uint32_t *textureBuffer;
             };
