+++ plugins/fengyun4_support/fy4/module_fy4_lrit_data_decoder.h
@@ -29,7 +29,7 @@ namespace fy4
 
                 // UI Stuff
                 bool hasToUpdate = false;
-                unsigned int textureID = 0;
+                intptr_t textureID = 0;
                 uint32_t *textureBuffer;
             };
