+++ plugins/gk2a_support/gk2a/module_gk2a_lrit_data_decoder.h
@@ -36,7 +36,7 @@ namespace gk2a
 
                 // UI Stuff
                 bool hasToUpdate = false;
-                unsigned int textureID = 0;
+                intptr_t textureID = 0;
                 uint32_t *textureBuffer;
             };
